moved trace sizes somewhere shareable
